backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
pydantic
openai-whisper
yt-dlp
ffmpeg-python